memtune: Let virsh know the unlimited value for memory tunables
Display or set unlimited values for memory parameters. Unlimited is
represented by INT64_MAX in memory cgroup.

Reported-by: Justin Clift <jclift@redhat.com>
(cherry picked from commit d94a14f89d2e421597c3d38e5e5e1c613629af06)

+++ src/qemu/qemu_driver.c
@@ -7519,7 +7519,7 @@ static int qemuDomainGetMemoryParameters(virDomainPtr dom,
     int i;
     virCgroupPtr group = NULL;
     virDomainObjPtr vm = NULL;
-    unsigned long val;
+    unsigned long long val;
     int ret = -1;
     int rc;
